custom error instance

+++ src/storage/group/groupCreate.ts
@@ -1,5 +1,6 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { GROUP_COLLECTION } from '@storage/storageConfig';
+import { AppError } from '@utils/AppError';
 import { groupsGetAll } from './groupsGetAll';
 
 export async function groupCreate(newGroup: string) {
@@ -7,6 +8,12 @@ export async function groupCreate(newGroup: string) {
 		const storedGroups = await groupsGetAll();
 
 		const storage = JSON.stringify([...storedGroups, newGroup]);
+
+		const groupAlreadyExists = storedGroups.includes(newGroup);
+
+		if (groupAlreadyExists) {
+			throw new AppError('Já existe um grupo cadastrado com esse nome.');
+		}
 
 		await AsyncStorage.setItem(GROUP_COLLECTION, storage);
 	} catch (error) {
